Chore: Moved UserInfo from foundation.ts to compas/foundation.ts, since it's compas specific
