+++ src/foundation.ts
@@ -479,23 +479,6 @@ export function newOpenDocEvent(
   });
 }
 
-/** Represents user information from a backend. */
-export interface UserInfoDetail {
-  name: string;
-}
-export type UserInfoEvent = CustomEvent<UserInfoDetail>;
-export function newUserInfoEvent(
-  name: string,
-  eventInitDict?: CustomEventInit<Partial<UserInfoDetail>>
-): UserInfoEvent {
-  return new CustomEvent<UserInfoDetail>('userinfo', {
-    bubbles: true,
-    composed: true,
-    ...eventInitDict,
-    detail: { name, ...eventInitDict?.detail },
-  });
-}
-
 /** @returns a reference to `element` with segments delimited by '/'. */
 // TODO(c-dinkel): replace with identity (FIXME)
 export function referencePath(element: Element): string {
@@ -2880,7 +2863,6 @@ declare global {
     ['pending-state']: PendingStateEvent;
     ['editor-action']: EditorActionEvent<EditorAction>;
     ['open-doc']: OpenDocEvent;
-    ['userinfo']: UserInfoEvent;
     ['wizard']: WizardEvent;
     ['validate']: ValidateEvent;
     ['log']: LogEvent;
